discord_bot/requirements.txt
discord.py==2.4.0
firebase-admin==6.5.0
python-dotenv==1.0.1

